// sanity.types.ts
export type SiteContent = {
  _id: string;
  _type: "siteContent";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  landingPage?: {
    hero?: {
      prefix?: string;
      highlightedText?: string;
      description?: string;
    };
    smarterTools?: {
      title?: string;
      description?: string;
      tools?: Array<{
        title?: string;
        description?: string;
        icon?: "/feature-1.svg" | "/feature-2.svg" | "/feature-3.svg" | "/feature-4.svg" | "/feature-5.svg";
        color?: "#AB2DB5" | "#148BD0" | "#80A71B" | "#F6B40C" | "#5E24D0";
        _key: string;
      }>;
    };
    howItWorks?: {
      title?: string;
      steps?: Array<{
        title?: string;
        description?: string;
        _key: string;
      }>;
    };
    whatToAchieve?: {
      title?: string;
      suffix?: string;
      suffix2?: string;
      subtitle?: string;
      options?: Array<{
        title?: string;
        subtitle?: string;
        icon?: "/landing/achieve-1.svg" | "/landing/achieve-2.svg" | "/landing/achieve-3.svg" | "/landing/achieve-4.svg" | "/landing/achieve-5.svg";
        _key: string;
      }>;
    };
    moneyStress?: {
      title?: string;
      description?: Array<{
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?: "normal" | "h1" | "h2" | "h3" | "h4" | "blockquote";
        listItem?: "bullet";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      } | {
        asset?: {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
        };
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        alt?: string;
        _type: "image";
        _key: string;
      }>;
    };
    yourMoney?: {
      title?: string;
      description?: string;
      options?: Array<{
        title?: string;
        description?: string;
        icon?: "/feature2-1.svg" | "/feature2-2.svg" | "/feature2-3.svg" | "/feature2-4.svg";
        _key: string;
      }>;
    };
    faq?: {
      title?: string;
      questions?: Array<{
        question?: string;
        answer?: string;
        _key: string;
      }>;
    };
  };
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;